requirements.txt
# requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
line-bot-sdk==3.12.0
anthropic>=0.45.0
httpx>=0.27.0
sentence-transformers==3.1.1
faiss-cpu==1.8.0
pandas==2.2.3
numpy==1.26.4
python-dotenv==1.0.1
